requirements.txt
escpos==2.0.0
fpdf==1.7.2
pyserial==3.5
PySide6==6.8.1.1
python_dateutil==2.9.0
Requests==2.32.3
sportident==1.2.8
SQLAlchemy==2.0.37
PyYAML==6.0.2
Jinja2==3.1.6